src/ingestion/requirements.txt
boto3==1.28.57
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0